added ~/.sys directory
Added this directory to contain everything I may need to setup a system.
As of right now all I have is a notes directory with a single note that
explains how to get specific bluetooth devices to automatically connect
once the computer has awoken from suspend. The instructions may be
specific to pop-os but I'm not actually sure.