add padding to vte terminals via gtk css
Current terminal I'm using is tilix which handles creating splits
in a window but by removing the terminal title bar, the text of the
terminal was too close to the border. I add padding to the terminal by
editing gtk.css since it can be configured like any vte terminal.

Note that this is obviously gtk-specific.