assortment of changes, mainly theme and colors
I attempt to make a highlight theme that makes ranger show a preview to
match the vim theme as much as possible. It's still far from the actual
theme but it's good enough for now.

Some other changes, I found a way to customize colors specific to a
language. This is what colors.vim is meant to be used for. I need to
pull in any color related things into this file since there may be some
other spots I'm doing that.

I also added a which-key shortcut for highlighting words, mostly to help
memorize that it can just be done via "*".