add coc and other dev plugins, configs for neovim
This plugins added in this change are:

- vim-commentary
- coc
- vim-startify
- ack.vim

Most are fully good-to-go except coc. That requires more configuration
and in general, learning about how to work with it. One thing that is
becoming obvious is the need to clean up my `Leader` key based shortcuts
and rethink how to organize. A lot of them are repeats of ones with
shortcuts that I wonder if I really need them.