remove coc autocomplete first result on CR
This removes the feature in nvim coc where if the autocomplete menu is
displayed and carriage return (enter) is pressed, then the first item on
the menu is automatically selected. This doesn't work well because of
the situation where the first item is not what you want and what you're
typing is the last thing you want on that line so you instinctively
press enter. This would incorrectly select that bad suggestion.

The instance this was occurring a lot was in elixir when typing `do`.
It's very common for it to be the last term and coc was always
converting it to `doc` because I would press enter.