switch to fnm, tune to work on macs
This change swaps out nvm for fnm to manage node versions. Doing this
led to an issue for how to appropriately add to PATH in config.fish
(why is this so confusing?) which led me to a change made to fish to
include a function to easily add paths regardless of whether its done in
the terminal or config.fish. However this will only be released in
v3.2.0 where the current release is something like v3.1.2, so I just
copied the function here with the intention of deleting it once bumping
fish.

Also there were some adjustments to make this work better on macs,
mainly changing the path for nvim.